# NodeJS/node_mvc_rush/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "",
    "version": "2",
    "title": "View of MVC"
  },
  "host": "localhost:3000",
  "tags": [
    {
      "name": "AYE"
    }
  ],
  "schemes": [
    "http"
  ],
  "paths": {
    "/api/ayetest": {
      "post": {
        "tags": [
          "AYE"
        ],
        "summary": "testSWAGGER",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "value",
            "in": "formData",
            "description": "เลข input",
            "required": true,
            "type": "string",
            "default": "1"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/transaction200"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/badRequest400"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/transaction500"
            }
          }
        }
      }
    },
    "/api/LoginStudent": {
      "post": {
        "tags": [
          "StudentLogin"
        ],
        "summary": "การเข้าสู่ระบบของนักศึกษา",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "ID",
            "in": "formData",
            "description": "ไอดีนักศีกษา",
            "required": true,
            "type": "string",
            "default": "59050200"
          },
          {
            "name": "Pass",
            "in": "formData",
            "description": "รหัสผ่าน",
            "required": true,
            "type": "string",
            "default": "211240"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/transaction200"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/badRequest400"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/transaction500"
            }
          }
        }
      }
    },
    "/api/UpdateData": {
      "post": {
        "tags": [
          "StudentLogin"
        ],
        "summary": "เปลี่ยนสถานะการเช็คชื่อของนักศึกษา",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "ID",
            "in": "formData",
            "description": "ไอดีนักศีกษา",
            "required": true,
            "type": "string",
            "default": "59050200"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/transaction200"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/badRequest400"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/transaction500"
            }
          }
        }
      }
    },
    "/api/List": {
      "get": {
        "tags": [
          "StudentLogin"
        ],
        "summary": "นวนผู้ลงทะเบียนของนักศึกษา",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/transaction200"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/badRequest400"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/transaction500"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "badRequest400": {
      "type": "object",
      "properties": {
        "statusCode": {
          "type": "number",
          "example": 400
        },
        "message": {
          "type": "string",
          "example": "<function_name> account_id is required"
        }
      }
    },
    "transaction200": {
      "type": "object",
      "properties": {
        "statusCode": {
          "type": "number",
          "example": 200
        },
        "message": {
          "type": "string",
          "example": "Transaction has been submitted"
        }
      }
    },
    "transaction500": {
      "type": "object",
      "properties": {
        "statusCode": {
          "type": "number",
          "example": 500
        },
        "message": {
          "type": "string",
          "example": "Internal Server Error"
        }
      }
    }
  }
}